+++ rust/operator-binary/src/hbase_controller.rs
@@ -8,6 +8,7 @@ use std::{
 };
 
 use const_format::concatcp;
+use indoc::formatdoc;
 use product_config::{
     ProductConfigManager,
     types::PropertyNameKind,
@@ -20,8 +21,14 @@ use stackable_operator::{
         configmap::ConfigMapBuilder,
         meta::ObjectMetaBuilder,
         pod::{
-            PodBuilder, container::ContainerBuilder, resources::ResourceRequirementsBuilder,
+            PodBuilder,
+            container::ContainerBuilder,
+            resources::ResourceRequirementsBuilder,
             security::PodSecurityContextBuilder,
+            volume::{
+                ListenerOperatorVolumeSourceBuilder, ListenerOperatorVolumeSourceBuilderError,
+                ListenerReference,
+            },
         },
     },
     cluster_resources::{ClusterResourceApplyStrategy, ClusterResources},
@@ -70,9 +77,11 @@ use crate::{
         construct_role_specific_non_heap_jvm_args,
     },
     crd::{
-        APP_NAME, AnyServiceConfig, Container, HBASE_ENV_SH, HBASE_REST_PORT_NAME_HTTP,
-        HBASE_REST_PORT_NAME_HTTPS, HBASE_SITE_XML, HbaseClusterStatus, HbaseRole,
-        JVM_SECURITY_PROPERTIES_FILE, SSL_CLIENT_XML, SSL_SERVER_XML, merged_env, v1alpha1,
+        APP_NAME, AnyServiceConfig, Container, HBASE_ENV_SH, HBASE_MASTER_PORT,
+        HBASE_REGIONSERVER_PORT, HBASE_REST_PORT_NAME_HTTP, HBASE_REST_PORT_NAME_HTTPS,
+        HBASE_SITE_XML, HbaseClusterStatus, HbaseRole, JVM_SECURITY_PROPERTIES_FILE,
+        LISTENER_VOLUME_DIR, LISTENER_VOLUME_NAME, SSL_CLIENT_XML, SSL_SERVER_XML, merged_env,
+        v1alpha1,
     },
     discovery::build_discovery_configmap,
     kerberos::{
@@ -312,6 +321,16 @@ pub enum Error {
 
     #[snafu(display("failed to construct JVM arguments"))]
     ConstructJvmArgument { source: crate::config::jvm::Error },
+
+    #[snafu(display("failed to build Labels"))]
+    LabelBuild {
+        source: stackable_operator::kvp::LabelError,
+    },
+
+    #[snafu(display("failed to build listener volume"))]
+    BuildListenerVolume {
+        source: ListenerOperatorVolumeSourceBuilderError,
+    },
 }
 
 type Result<T, E = Error> = std::result::Result<T, E>;
@@ -375,26 +394,6 @@ pub async fn reconcile_hbase(
     )
     .context(CreateClusterResourcesSnafu)?;
 
-    let region_server_role_service =
-        build_region_server_role_service(hbase, &resolved_product_image)?;
-    cluster_resources
-        .add(client, region_server_role_service)
-        .await
-        .context(ApplyRoleServiceSnafu)?;
-
-    // discovery config map
-    let discovery_cm = build_discovery_configmap(
-        hbase,
-        &client.kubernetes_cluster_info,
-        &zookeeper_connection_information,
-        &resolved_product_image,
-    )
-    .context(BuildDiscoveryConfigMapSnafu)?;
-    cluster_resources
-        .add(client, discovery_cm)
-        .await
-        .context(ApplyDiscoveryConfigMapSnafu)?;
-
     let (rbac_sa, rbac_rolebinding) = build_rbac_resources(
         hbase,
         APP_NAME,
@@ -443,7 +442,6 @@ pub async fn reconcile_hbase(
             )?;
             let rg_statefulset = build_rolegroup_statefulset(
                 hbase,
-                &client.kubernetes_cluster_info,
                 &hbase_role,
                 &rolegroup,
                 rolegroup_config,
@@ -484,6 +482,20 @@ pub async fn reconcile_hbase(
         }
     }
 
+    // Discovery CM will fail to build until the rest of the cluster has been deployed, so do it last
+    // so that failure won't inhibit the rest of the cluster from booting up.
+    let discovery_cm = build_discovery_configmap(
+        hbase,
+        &client.kubernetes_cluster_info,
+        &zookeeper_connection_information,
+        &resolved_product_image,
+    )
+    .context(BuildDiscoveryConfigMapSnafu)?;
+    cluster_resources
+        .add(client, discovery_cm)
+        .await
+        .context(ApplyDiscoveryConfigMapSnafu)?;
+
     let cluster_operation_cond_builder =
         ClusterOperationsConditionBuilder::new(&hbase.spec.cluster_operation);
 
@@ -501,59 +513,6 @@ pub async fn reconcile_hbase(
         .context(ApplyStatusSnafu)?;
 
     Ok(Action::await_change())
-}
-
-/// The server-role service is the primary endpoint that should be used by clients that do not perform internal load balancing,
-/// including targets outside of the cluster.
-pub fn build_region_server_role_service(
-    hbase: &v1alpha1::HbaseCluster,
-    resolved_product_image: &ResolvedProductImage,
-) -> Result<Service> {
-    let role = HbaseRole::RegionServer;
-    let role_name = role.to_string();
-    let role_svc_name = hbase
-        .server_role_service_name()
-        .context(GlobalServiceNameNotFoundSnafu)?;
-    let ports = hbase
-        .ports(&role, &resolved_product_image.product_version)
-        .into_iter()
-        .map(|(name, value)| ServicePort {
-            name: Some(name),
-            port: i32::from(value),
-            protocol: Some("TCP".to_string()),
-            ..ServicePort::default()
-        })
-        .collect();
-
-    let metadata = ObjectMetaBuilder::new()
-        .name_and_namespace(hbase)
-        .name(&role_svc_name)
-        .ownerreference_from_resource(hbase, None, Some(true))
-        .context(ObjectMissingMetadataForOwnerRefSnafu)?
-        .with_recommended_labels(build_recommended_labels(
-            hbase,
-            &resolved_product_image.app_version_label,
-            &role_name,
-            "global",
-        ))
-        .context(ObjectMetaSnafu)?
-        .build();
-
-    let service_selector_labels =
-        Labels::role_selector(hbase, APP_NAME, &role_name).context(BuildLabelSnafu)?;
-
-    let service_spec = ServiceSpec {
-        type_: Some(hbase.spec.cluster_config.listener_class.k8s_service_type()),
-        ports: Some(ports),
-        selector: Some(service_selector_labels.into()),
-        ..ServiceSpec::default()
-    };
-
-    Ok(Service {
-        metadata,
-        spec: Some(service_spec),
-        status: None,
-    })
 }
 
 /// The rolegroup [`ConfigMap`] configures the rolegroup based on the configuration given by the administrator
@@ -589,6 +548,63 @@ fn build_rolegroup_config_map(
                 );
                 hbase_site_config
                     .extend(hbase_opa_config.map_or(vec![], |config| config.hbase_site_config()));
+
+                // Get more useful stack traces...
+                // The default Netty impl gives us netty garbage and nothing else
+                hbase_site_config.insert(
+                    "hbase.rpc.client.impl".to_string(),
+                    "org.apache.hadoop.hbase.ipc.BlockingRpcClient".to_string(),
+                );
+
+                // Set listener endpoint information with generic properties
+                hbase_site_config.insert(
+                    "hbase.listener.endpoint".to_string(),
+                    "${HBASE_LISTENER_ENDPOINT}".to_string(),
+                );
+                hbase_site_config.insert(
+                    "hbase.info.port".to_string(),
+                    "${HBASE_INFO_PORT}".to_string(),
+                );
+
+                match hbase_role {
+                    HbaseRole::Master => {
+                        hbase_site_config.insert(
+                            "hbase.master.ipc.address".to_string(),
+                            "0.0.0.0".to_string(),
+                        );
+                        hbase_site_config.insert(
+                            "hbase.master.ipc.port".to_string(),
+                            HBASE_MASTER_PORT.to_string(),
+                        );
+                        hbase_site_config.insert(
+                            "hbase.master.hostname".to_string(),
+                            "${HBASE_SERVICE_HOST}".to_string(),
+                        );
+                        hbase_site_config.insert(
+                            "hbase.master.port".to_string(),
+                            "${HBASE_SERVICE_PORT}".to_string(),
+                        );
+                    }
+                    HbaseRole::RegionServer => {
+                        hbase_site_config.insert(
+                            "hbase.regionserver.ipc.address".to_string(),
+                            "0.0.0.0".to_string(),
+                        );
+                        hbase_site_config.insert(
+                            "hbase.regionserver.ipc.port".to_string(),
+                            HBASE_REGIONSERVER_PORT.to_string(),
+                        );
+                        hbase_site_config.insert(
+                            "hbase.unsafe.regionserver.hostname".to_string(),
+                            "${HBASE_SERVICE_HOST}".to_string(),
+                        );
+                        hbase_site_config.insert(
+                            "hbase.regionserver.port".to_string(),
+                            "${HBASE_SERVICE_PORT}".to_string(),
+                        );
+                    }
+                    HbaseRole::RestServer => {}
+                };
 
                 // configOverride come last
                 hbase_site_config.extend(config.clone());
@@ -732,7 +748,7 @@ fn build_rolegroup_service(
 
     let metadata = ObjectMetaBuilder::new()
         .name_and_namespace(hbase)
-        .name(rolegroup.object_name())
+        .name(format!("{name}-metrics", name = rolegroup.object_name()))
         .ownerreference_from_resource(hbase, None, Some(true))
         .context(ObjectMissingMetadataForOwnerRefSnafu)?
         .with_recommended_labels(build_recommended_labels(
@@ -772,7 +788,6 @@ fn build_rolegroup_service(
 #[allow(clippy::too_many_arguments)]
 fn build_rolegroup_statefulset(
     hbase: &v1alpha1::HbaseCluster,
-    cluster_info: &KubernetesClusterInfo,
     hbase_role: &HbaseRole,
     rolegroup_ref: &RoleGroupRef<v1alpha1::HbaseCluster>,
     rolegroup_config: &HashMap<PropertyNameKind, BTreeMap<String, String>>,
@@ -867,15 +882,30 @@ fn build_rolegroup_statefulset(
         },
     ]);
 
+    let role_name = hbase_role.cli_role_name();
     let mut hbase_container = ContainerBuilder::new("hbase").expect("ContainerBuilder not created");
+
+    let rest_http_port_name = if hbase.has_https_enabled() {
+        HBASE_REST_PORT_NAME_HTTPS
+    } else {
+        HBASE_REST_PORT_NAME_HTTP
+    };
+
     hbase_container
         .image_from_product_image(resolved_product_image)
-        .command(vec!["/stackable/hbase/bin/hbase-entrypoint.sh".to_string()])
-        .args(vec![
-            hbase_role.cli_role_name(),
-            hbase_service_domain_name(hbase, rolegroup_ref, cluster_info)?,
-            hbase.service_port(hbase_role).to_string(),
-        ])
+        .command(command())
+        .args(vec![formatdoc! {"
+            {entrypoint} {role} {port} {port_name} {ui_port_name}",
+            entrypoint = "/stackable/hbase/bin/hbase-entrypoint.sh".to_string(),
+            role = role_name,
+            port = hbase.service_port(hbase_role).to_string(),
+            port_name = match hbase_role {
+                HbaseRole::Master => "master",
+                HbaseRole::RegionServer => "regionserver",
+                HbaseRole::RestServer => rest_http_port_name,
+            },
+            ui_port_name = hbase.ui_port_name(),
+        }])
         .add_env_vars(merged_env)
         // Needed for the `containerdebug` process to log it's tracing information to.
         .add_env_var(
@@ -890,6 +920,8 @@ fn build_rolegroup_statefulset(
         .context(AddVolumeMountSnafu)?
         .add_volume_mount("log", STACKABLE_LOG_DIR)
         .context(AddVolumeMountSnafu)?
+        .add_volume_mount(LISTENER_VOLUME_NAME, LISTENER_VOLUME_DIR)
+        .context(AddVolumeMountSnafu)?
         .add_container_ports(ports)
         .resources(merged_config.resources().clone().into())
         .startup_probe(startup_probe)
@@ -898,13 +930,17 @@ fn build_rolegroup_statefulset(
 
     let mut pod_builder = PodBuilder::new();
 
+    let recommended_object_labels = build_recommended_labels(
+        hbase,
+        hbase_version,
+        &rolegroup_ref.role,
+        &rolegroup_ref.role_group,
+    );
+    let recommended_labels =
+        Labels::recommended(recommended_object_labels.clone()).context(LabelBuildSnafu)?;
+
     let pb_metadata = ObjectMetaBuilder::new()
-        .with_recommended_labels(build_recommended_labels(
-            hbase,
-            hbase_version,
-            &rolegroup_ref.role,
-            &rolegroup_ref.role_group,
-        ))
+        .with_recommended_labels(recommended_object_labels)
         .context(ObjectMetaSnafu)?
         .build();
 
@@ -939,6 +975,15 @@ fn build_rolegroup_statefulset(
         .context(AddVolumeSnafu)?
         .service_account_name(service_account.name_any())
         .security_context(PodSecurityContextBuilder::new().fs_group(1000).build());
+
+    // externally-reachable listener endpoints should use a pvc volume...
+    let pvc = ListenerOperatorVolumeSourceBuilder::new(
+        &ListenerReference::ListenerClass(merged_config.listener_class().to_string()),
+        &recommended_labels,
+    )
+    .context(BuildListenerVolumeSnafu)?
+    .build_pvc(LISTENER_VOLUME_NAME.to_string())
+    .context(BuildListenerVolumeSnafu)?;
 
     if let Some(ContainerLogConfig {
         choice:
@@ -1042,8 +1087,12 @@ fn build_rolegroup_statefulset(
             match_labels: Some(statefulset_match_labels.into()),
             ..LabelSelector::default()
         },
-        service_name: Some(rolegroup_ref.object_name()),
+        service_name: Some(format!(
+            "{name}-metrics",
+            name = rolegroup_ref.object_name()
+        )),
         template: pod_template,
+        volume_claim_templates: Some(vec![pvc]),
         ..StatefulSetSpec::default()
     };
 
@@ -1052,6 +1101,17 @@ fn build_rolegroup_statefulset(
         spec: Some(statefulset_spec),
         status: None,
     })
+}
+
+/// Returns the container command.
+fn command() -> Vec<String> {
+    vec![
+        "/bin/bash".to_string(),
+        "-x".to_string(),
+        "-euo".to_string(),
+        "pipefail".to_string(),
+        "-c".to_string(),
+    ]
 }
 
 fn write_hbase_env_sh<'a, T>(properties: T) -> String
@@ -1116,6 +1176,7 @@ fn build_hbase_env_sh(
     let role_specific_non_heap_jvm_args =
         construct_role_specific_non_heap_jvm_args(hbase, hbase_role, role_group, product_version)
             .context(ConstructJvmArgumentSnafu)?;
+
     match hbase_role {
         HbaseRole::Master => {
             result.insert(
@@ -1156,28 +1217,6 @@ fn validate_cr(hbase: &v1alpha1::HbaseCluster) -> Result<()> {
     }
     tracing::info!("End CR validation");
     Ok(())
-}
-
-/// Build the domain name of an HBase service pod.
-/// The hbase-entrypoint.sh script uses this to build the fully qualified name of a pod
-/// by appending it to the `HOSTNAME` environment variable.
-/// This name is required by the RegionMover to function properly.
-fn hbase_service_domain_name(
-    hbase: &v1alpha1::HbaseCluster,
-    rolegroup_ref: &RoleGroupRef<v1alpha1::HbaseCluster>,
-    cluster_info: &KubernetesClusterInfo,
-) -> Result<String, Error> {
-    let hbase_cluster_name = rolegroup_ref.object_name();
-    let pod_namespace = hbase
-        .metadata
-        .namespace
-        .clone()
-        .context(ObjectHasNoNamespaceSnafu)?;
-    let cluster_domain = &cluster_info.cluster_domain;
-
-    Ok(format!(
-        "{hbase_cluster_name}.{pod_namespace}.svc.{cluster_domain}"
-    ))
 }
 
 #[cfg(test)]